FIX: Header 프로필 메뉴 컴포넌트 네이밍 변경 및 스타일 변경
