--- frontend/src/components/common/Header/HeaderOption/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./HeaderOption";
