+++ frontend/src/components/common/HeaderProfileMenu/HeaderProfileMenu.scss
@@ -1,7 +1,6 @@
-.header-option {
-  top: 3rem;
-  left: 0rem;
-  right: 0;
+.header-profile-menu {
+  top: 3.2rem;
+  left: 3.4rem;
   z-index: 1000;
   display: flex;
   flex-direction: column;
@@ -11,6 +10,31 @@
   height: auto;
   position: absolute;
   border-radius: 0.2rem;
+
+  &:after,
+  &:before {
+    bottom: 100%;
+    left: 25%;
+    border: solid transparent;
+    content: "";
+    height: 0;
+    width: 0;
+    position: absolute;
+    pointer-events: none;
+  }
+
+  &:after {
+    border-color: rgba(163, 163, 163, 0);
+    border-bottom-color: var(--theme-main-background);
+    border-width: 10px;
+    margin-left: -10px;
+  }
+  &:before {
+    border-color: rgba(162, 162, 162, 0);
+    border-bottom-color: var(--theme-border);
+    border-width: 12px;
+    margin-left: -12px;
+  }
 
   &-logout {
     &-wrap {
