+++ frontend/src/components/common/HeaderProfileMenu/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./HeaderProfileMenu";
